refactor: 총상금 타입 변경

+++ src/main/java/lotto/domain/LottoResult.java
@@ -18,9 +18,9 @@ public class LottoResult implements Rate{
                 .count();
     }
 
-    int totalPrize() {
+    long totalPrize() {
         return result.stream()
-                .mapToInt(prize -> prize.getPrize())
+                .mapToLong(prize -> prize.getPrize())
                 .sum();
     }
